chore(frontend): setup provider dir structure
